+++ R/udunits.R
@@ -1,3 +1,16 @@
+#' Test if two units are convertible
+#'
+#' @param u1 string, the first unit
+#' @param u2 string, the second unit
+#'
+#' @return boolean, \code{TRUE} if both units are convertible.
+#' @export
+#'
+#' @examples
+#' ud_are_convertible("m", "km")
+#' a <- set_units(1:3, m/s)
+#' ud_are_convertible(units(a), "km/h)
+#' ud_are_convertible("s", "kg")
 ud_are_convertible = function(u1, u2) {
 	res <- try(R_ut_are_convertible(R_ut_parse(as.character(u1)), R_ut_parse(as.character(u2))),
 		silent = TRUE)